fix: cultivate plan jumps multiple times
