+++ app/(tabs)/toolbox.tsx
@@ -1,14 +1,6 @@
-import { useRouter, type Href, type Router } from 'expo-router';
+import { useRouter, type Router } from 'expo-router';
 import { useEffect, useState } from 'react';
-import { Alert, FlatList, Linking } from 'react-native';
-import { toast } from 'sonner-native';
-
-import Banner, { type BannerContent } from '@/components/banner';
-import PageContainer from '@/components/page-container';
-import { Button } from '@/components/ui/button';
-import { Text } from '@/components/ui/text';
-
-import { pushToWebViewJWCH, pushToWebViewNormal } from '@/lib/webview';
+import { Alert, FlatList } from 'react-native';
 
 import BannerImage1 from '@/assets/images/banner/default_banner1.webp';
 import BannerImage2 from '@/assets/images/banner/default_banner2.webp';
@@ -26,35 +18,15 @@ import IDCardIcon from '@/assets/images/toolbox/ic_studentcard.svg';
 import WikiIcon from '@/assets/images/toolbox/ic_wiki.svg';
 import XuankeIcon from '@/assets/images/toolbox/ic_xuanke.svg';
 import ZHCTIcon from '@/assets/images/toolbox/ic_zhct.svg';
+import Banner, { type BannerContent } from '@/components/banner';
+import PageContainer from '@/components/page-container';
+import { Button } from '@/components/ui/button';
+import { Text } from '@/components/ui/text';
+
+import { pushToWebViewJWCH, pushToWebViewNormal } from '@/lib/webview';
+import { ToolType, toolOnPress, type Tool } from '@/utils/tools';
 
 // 工具类型的枚举
-enum ToolType {
-  LINK = 'link', // 跳转路由
-  URL = 'URL', // 打开网页
-  FUNCTION = 'function', // 执行函数
-  NULL = 'null', // 空操作
-}
-
-type Tool = {
-  name: string;
-  icon: any;
-} & (
-  | {
-      type: ToolType.LINK;
-      href: Href;
-    }
-  | {
-      type: ToolType.URL;
-      href: string;
-    }
-  | {
-      type: ToolType.FUNCTION;
-      action: (router: ReturnType<typeof useRouter>) => void | Promise<void>;
-    }
-  | {
-      type: ToolType.NULL;
-    }
-);
 
 // 常量：横幅数据
 const DEFAULT_BANNERS: BannerContent[] = [
@@ -183,26 +155,6 @@ const useToolsPageData = () => {
   }, []);
 
   return { bannerList, toolList };
-};
-
-// 工具按钮的点击事件
-const toolOnPress = (tool: Tool, router: ReturnType<typeof useRouter>) => {
-  switch (tool.type) {
-    case ToolType.NULL: // 空操作
-      break;
-    case ToolType.LINK: // 跳转路由
-      router.push(tool.href);
-      break;
-    case ToolType.URL: // 打开网页
-      Linking.openURL(tool.href).catch(err => Alert.alert('错误', '无法打开链接 (' + err + ')'));
-      break;
-    case ToolType.FUNCTION: // 执行函数，并传入 router 参数
-      tool.action(router);
-      break;
-    default:
-      toast.error('未知工具类型');
-      console.error('未知工具类型', tool);
-  }
 };
 
 // 工具按钮的渲染函数
